+++ CHANGELOG.md
@@ -5,9 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Feature] Add isFilter prop to <Input> for search icon (#488)
+### Changed
+- [Patch] Update disclose arrow to narrow style per Silvia. (#504)
 
 ## [14.0.5][14.0.5] - 2016-08-17
 ### Added
+- [Feature] Add isFilter prop to <Input> for search icon (#488)
 - [Patch] Update require.css statement to ignore local webpack config transforms. Minor version bump from publish issues (user error). (#498)
 
 ## [14.0.0][14.0.0] - 2016-08-15