+++ CHANGELOG.md
@@ -5,12 +5,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Feature] Add isFilter prop to <Input> for search icon (#488)
+
+### Changed
+- [Patch] Update disclose arrow to narrow style per Silvia. (#504)
+- [Patch] Fix typo in React prop description to say onChange fires every keypress. (#492)
+- [Patch] Fix disabled button namespace with OUI. Add all disabled examples to React docs for easy testing for now. (#511)
+- [Patch] Typo for <Input> React prop where onChange triggers on every keypress. (#510)
 
 ## [14.0.3][14.0.3] - 2016-08-17
 ### Fixed
 - [Patch] Update require.css statement to ignore local webpack config transforms. Version patch bump due to publish user error. (#498)
 ## [14.0.4][14.0.4] - 2016-08-17
+## [14.0.5][14.0.5] - 2016-08-17
 ### Added
+- [Feature] Add isFilter prop to <Input> for search icon (#488)
 - [Patch] Update require.css statement to ignore local webpack config transforms. Minor version bump from publish issues (user error). (#498)
 
 ## [14.0.0][14.0.0] - 2016-08-15
@@ -525,3 +535,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [14.0.0]: https://github.com/optimizely/oui/compare/v13.2.0...v14.0.0
 [14.0.3]: https://github.com/optimizely/oui/compare/v14.0.0...v14.0.3
 [14.0.4]: https://github.com/optimizely/oui/compare/v14.0.0...v14.0.4
+[14.0.5]: https://github.com/optimizely/oui/compare/v14.0.0...v14.0.5